bug introduced with time optimisation fixed
Sometimes, with some trajectory (eg. traj with RNA/DNA) if you want to
clusterize the DNA/RNA and align to the protein backbone, it fails
because MDTRAJ didn't recognize DNA/RNA backbone.
Now the alignement is made before the extraction of the trajectory.
If you don't want protein alignement, use "-sr none"